Make the PKParser.parse call obey self.startRuleName when it comes to starting the parse.
This allows the caller to specify a non-default startRuleName before
parsing so that the parse starts with the specified non-terminal
(e.g. for testing).

This is implemented using a new template PGDynamicMethodCallTemplate.txt.
This calls the selector dynamically if its name has been specified,
otherwise it falls back to a default call.